Remove executable flag where not warranted
The executable flag is used to determine whether a file should be able
to be executed in a shell. Git records the executable bit, and ensures
its set on files when they're checked out.

The executable flag seems to be set on every file in this repository.
Usually, this is the result of developing on what Git calls a "broken
filesystem" (with respect to the executable bit) - most likely FAT.

See the core.fileMode setting to stop Git from marking every file
executable: http://stackoverflow.com/questions/1580596/how-do-i-make-git-ignore-mode-changes-chmod

This commit marks the vast majority of the files in the repo
non-executable, as they should be. Some files (like the /symfony task
script) retain the executable bit.